work around eclipse behavior..
paste method cant be accessed by reflection through its own class, because ScriptConsoleStyledText was private) because of that CTRL-V didnt work. Because that tries to call ScriptConsoleStyledText.paste() but that method is in a private class so not accessible.

This is ofcourse a bit weird. some how eclipse should set the method to accessible or look if it is accessible and not try to find an accessble method from a super class and then invoke it on the real component instance.